requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
bleach==6.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
diffusers==0.32.2
dnspython==2.8.0
fastapi==0.115.8
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.2.0
gensim==3.8.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.29.1
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.3.1
kaggle==1.7.4.5
kombu==5.6.2
MarkupSafe==3.0.2
mpmath==1.3.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.7
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==6.33.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytube==15.0.0
pytz==2025.1
PyYAML==6.0.2
redis==7.1.0
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.5.2
scipy==1.13.1
setuptools==69.5.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
text-unidecode==1.3
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.3
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xformers==0.0.29.post3
youtube-dl==2021.12.17
yt-dlp==2025.11.12
zipp==3.21.0